Add 22/04/2020 to timeline and hardcode values as int

+++ data/greece/isMOOD/append_json_to_csv.py
@@ -19,7 +19,7 @@ def append_json_to_csv():
     if today in time_series_df.columns:
         time_series_df = time_series_df.drop(today, axis =1)
     time_series_df = time_series_df.merge(new_data, left_on = 'Region', right_on = 'region_en_name', how = 'left')
-    
+    time_series_df.iloc[:,2:] = time_series_df.iloc[:,2:].astype('Int64')
     return time_series_df.to_csv(CASES_BY_REGION_TIME_SERIES_CSV_PATH, index = False)
